Server and Client
1. Connecting sockets:
For Client:
Create an object of the Socket class, specifying the server's IP address and port number.
For Server:
Create an object of the ServerSocket class, to wait for connections on the specified port.
Accept the client's connection request, thereby establishing a socket connection between the client and the server.
2. File upload from client to server:

At first Client:
Read all data from the file into a byte array.
Create an output stream via socket.
Write the byte array to the output stream, which will send the file data to the server.
And Server:
Create an input stream via socket.
Read the incoming byte array from the input stream.
Write the byte array to a file on the server's disk.
3. For File download from server to client:

Client:
Send a request for the specified file to the server.
Create an input stream via socket.
Read the incoming byte array from the input stream.
Write the byte array as a file to the client's disk.
Server:
Read the requested file into a byte array.
Create an output stream via socket.
Write the byte array to the output stream, which will send the file data to the client.